+++ src/entry-client.js
@@ -1,26 +1,27 @@
-import 'bootstrap/dist/css/bootstrap.min.css';
-import $ from 'jquery';
+import $ from './3rd/jquery';
+import 'bootstrap/dist/js/bootstrap';
+
 import './assets/prettify-jsdoc.css';
 import './assets/prettify-tomorrow.css';
 import './assets/tui-doc.less';
 import { createApp } from './app';
-import { showLnbApi, showLnbExamples, prettyPrint } from './mount';
+import { prettyPrint } from './mount';
 
 const { store } = createApp({}, { env: {} });
 
 prettyPrint();
-if (!store.state.isTutorial) {
-  const id = `${store.state.docs[0].longname}_sub`.replace(/"/g, '_');
-  var selectedApi = document.getElementById(id); // do not use jquery selector
-  var $selectedApi = $(selectedApi);
 
-  $selectedApi.removeClass('hidden');
-  $selectedApi
-    .parent()
-    .find('.glyphicon')
-    .removeClass('glyphicon-plus')
-    .addClass('glyphicon-minus');
-  showLnbApi();
+const id = `${store.state.docs[0].longname}_sub`.replace(/"/g, '_').replace(/[\\/]/g, '-');
+var selectedApi = document.getElementById(id); // do not use jquery selector
+if (selectedApi) {
+  var $selectedApi = $(selectedApi);
+  var $selectTab = $(`#${$selectedApi.parents('.lnb-api').data('member-tab')}`);
+  $selectedApi.parents('.collapse').collapse('show');
+  $selectedApi.collapse('show');
+  $selectTab.tab('show');
 } else {
-  showLnbExamples();
+  $('.lnb-tab')
+    .find('[data-toggle="tab"]')
+    .first()
+    .tab('show');
 }